refactor: Move `copyAttributeList(…)` to `living/node.js`
