+++ lib/jsdom/living/attributes.js
@@ -240,21 +240,6 @@ exports.removeAttributeByNameNS = function (element, namespace, localName) {
   return attr;
 };
 
-exports.copyAttributeList = function (sourceElement, destElement) {
-  // Needed by https://dom.spec.whatwg.org/#concept-node-clone
-
-  for (const sourceAttr of sourceElement._attributeList) {
-    const destAttr = attrGenerated.createImpl(destElement._globalObject, [], {
-      namespace: sourceAttr._namespace,
-      namespacePrefix: sourceAttr._namespacePrefix,
-      localName: sourceAttr._localName,
-      value: sourceAttr._value
-    });
-
-    exports.appendAttribute(destElement, destAttr);
-  }
-};
-
 exports.attributeNames = function (element) {
   // Needed by https://dom.spec.whatwg.org/#dom-element-getattributenames
