+++ lib/jsdom/living/node.js
@@ -1,5 +1,6 @@
 "use strict";
 const attributes = require("./attributes");
+const attrGenerated = require("./generated/Attr.js");
 const { cloningSteps, domSymbolTree } = require("./helpers/internal-constants");
 const NODE_TYPE = require("./node-type");
 const orderedSetParse = require("./helpers/ordered-set").parse;
@@ -31,7 +32,7 @@ module.exports.clone = function (node, document, cloneChildren) {
     case NODE_TYPE.ELEMENT_NODE:
       copy = document._createElementWithCorrectElementInterface(node._localName, node._namespaceURI);
       copy._prefix = node._prefix;
-      attributes.copyAttributeList(node, copy);
+      copyAttributeList(node, copy);
       break;
 
     case NODE_TYPE.TEXT_NODE:
@@ -68,6 +69,21 @@ module.exports.clone = function (node, document, cloneChildren) {
 
   return copy;
 };
+
+function copyAttributeList(sourceElement, destElement) {
+  // Needed by https://dom.spec.whatwg.org/#concept-node-clone
+
+  for (const sourceAttr of sourceElement._attributeList) {
+    const destAttr = attrGenerated.createImpl(destElement._globalObject, [], {
+      namespace: sourceAttr._namespace,
+      namespacePrefix: sourceAttr._namespacePrefix,
+      localName: sourceAttr._localName,
+      value: sourceAttr._value
+    });
+
+    attributes.appendAttribute(destElement, destAttr);
+  }
+}
 
 // For the following, memoization is not applied here since the memoized results are stored on `this`.
